"Refactor imports and remove unused components"
This commit removes the `/components/container/Details` directory and moves the `BookDetailsContainer` and `CategoryDetailsContainer` components to `/container/Details`. It also updates the imports in the related files. Additionally, it deletes the `HomeContainer` component from `/components/container/Home` and moves it to `/container/Home`. The `BookDetails` and `CategoryDetails` components in `/components/container/Details` are no longer used and have been removed.